Enhance enddate management for visits

+++ dbmodel/utils/ftrg/gw_trg_om_visit.sql
@@ -1,41 +1,60 @@
 ﻿
-CREATE OR REPLACE FUNCTION "SCHEMA_NAME".gw_trg_om_visit()
+CREATE OR REPLACE FUNCTION ws.gw_trg_om_visit()
   RETURNS trigger AS
 $BODY$
 DECLARE 
+
+v_version text;
   
 
 BEGIN
 
-    EXECUTE 'SET search_path TO '||quote_literal(TG_TABLE_SCHEMA)||', public';
+   EXECUTE 'SET search_path TO '||quote_literal(TG_TABLE_SCHEMA)||', public';
 
+   v_version = (SELECT giswater FROM ws.version ORDER by 1 desc LIMIT 1);
 
-    IF TG_OP='INSERT' THEN
+   IF TG_OP='INSERT' THEN
+	
+		-- automatic creation of workcat
+		IF (SELECT (value::json->>'AutoNewWorkcat') FROM config_param_system WHERE parameter='om_visit_parameters') THEN
+			INSERT INTO cat_work (id) VALUES (NEW.id);
+		END IF;
 
-	-- automatic creation of workcat
-	IF (SELECT (value::json->>'AutoNewWorkcat') FROM config_param_system WHERE parameter='om_visit_parameters') THEN
-		INSERT INTO cat_work (id) VALUES (NEW.id);
-	END IF;
+		-- setting values of enddate
+		IF NEW.is_done IS FALSE THEN
+			NEW.enddate=null;
+		END IF;	
 
-	RETURN NEW;
+		RETURN NEW;
 
 
     ELSIF TG_OP='UPDATE' THEN	
 
-	IF NEW.class_id != OLD.class_id THEN
-		DELETE FROM om_visit_event WHERE visit_id=NEW.id;			
-	END IF;
+		-- setting values of enddate
+		IF NEW.is_done IS FALSE THEN
+			NEW.enddate=null;
+		END IF;	
 
-	RETURN NEW;
+		IF v_version > '3.2.019' THEN
+			IF NEW.class_id != OLD.class_id THEN
+				DELETE FROM om_visit_event WHERE visit_id=NEW.id;			
+			END IF;
+		END IF;
+
+		RETURN NEW;
 				
     ELSIF TG_OP='DELETE' THEN
-	DELETE FROM cat_work WHERE id=OLD.id::text;
-	RETURN OLD;
+	
+    	-- automatic creation of workcat
+		IF (SELECT (value::json->>'AutoNewWorkcat') FROM config_param_system WHERE parameter='om_visit_parameters') THEN
+			DELETE FROM cat_work WHERE id=OLD.id::text;
+		END IF;
+		
+		RETURN OLD;
 		
     END IF;
-		
-END; 
+
+END;
 $BODY$
   LANGUAGE plpgsql VOLATILE
   COST 100;
-
